:sparkles: #57 좋아요 컬럼 수정 및 좋아요 관계 테이블/Entity 생성
