+++ src/main/java/com/example/SEENEMA/post/theater/domain/TheaterPost.java
@@ -63,7 +63,7 @@ public class TheaterPost {
     private LocalDateTime editedAt; // 최종 수정 일시
 
     @Column
-    private Boolean great;
+    private Long likeCount;
 
     @Builder
     public TheaterPost(User user, Theater theater, String title, String content, LocalDateTime createdAt, List<Tag> tags, List<Image> image) {
@@ -75,6 +75,6 @@ public class TheaterPost {
         this.tags=tags;
         this.viewCount= 0L;
         this.image = image;
-        this.great = Boolean.FALSE;
+        this.likeCount = 0L;
     }
 }
