+++ src/main/java/com/example/SEENEMA/post/theater/domain/TheaterPostHeart.java
@@ -1,0 +1,30 @@
+package com.example.SEENEMA.post.theater.domain;
+
+import com.example.SEENEMA.user.domain.User;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+import javax.persistence.*;
+
+@Getter
+@Setter
+@Table(name = "POST_THEATER_HEART")
+@NoArgsConstructor(access= AccessLevel.PROTECTED)
+@EntityListeners(AuditingEntityListener.class)
+@Entity
+public class TheaterPostHeart {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id")
+    private User user;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "post_no")
+    private TheaterPost theaterPost;
+}
