+++ src/main/java/com/example/SEENEMA/post/theater/dto/TheaterPostDto.java
@@ -52,7 +52,7 @@ public class TheaterPostDto {
         private List<Tag> tags;     // 게시글 태그
         private List<Image> image;
         private Long viewCount;    // 조회수
-        private Boolean great;      // 좋아요 여부
+        private Long likeCount;          // 좋아요 갯수
         private List<CommentDto.readComment> comments; // 댓글 목록
 
 
@@ -66,7 +66,7 @@ public class TheaterPostDto {
             this.editedAt = theaterPost.getEditedAt().toString();
             this.tags = theaterPost.getTags();
             this.viewCount = theaterPost.getViewCount();
-            this.great = theaterPost.getGreat();
+            this.likeCount = theaterPost.getLikeCount();
             this.image = new ArrayList<>(theaterPost.getImage());
         }
     }
